@@ -8,26 +8,27 @@
         "production": "mix --production"
     },
     "devDependencies": {
-        "axios": "^0.24.0",
+        "axios": "^0.26.1",
         "bootstrap": "^5.1.3",
         "cross-env": "^7.0.3",
         "jquery": "^3.6.0",
-        "laravel-mix": "^6.0.37",
+        "laravel-mix": "^6.0.43",
         "lodash": "^4.17.21",
         "popper.js": "^1.16",
-        "resolve-url-loader": "^4.0.0",
-        "sass": "^1.43.4",
-        "sass-loader": "^12.3.0",
-        "vue": "^2.6.14",
-        "vue-loader": "^15.9.8",
+        "resolve-url-loader": "^5.0.0",
+        "sass": "^1.49.9",
+        "sass-loader": "^12.6.0",
+        "vue": "^3.2.31",
+        "vue-loader": "^17.0.0",
         "vue-template-compiler": "^2.6.14"
     },
     "dependencies": {
+        "@suadelabs/vue3-multiselect": "^1.0.2",
+        "@vueform/toggle": "^2.0.1",
         "bootstrap-social": "^5.1.1",
-        "npm-check-updates": "^11.8.5",
+        "npm-check-updates": "^12.5.3",
         "vue-moment": "^4.1.0",
-        "vue-multiselect": "^2.1.6",
-        "vue-router": "^3.5.3",
-        "vuex": "^3.6.2"
+        "vue-router": "^4.0.14",
+        "vuex": "^4.0.2"
     }
 }
