@@ -7,7 +7,9 @@ use Illuminate\Notifications\Notifiable;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-use Laravel\Passport\HasApiTokens;
+use Laravel\Sanctum\HasApiTokens;
+
+//use Laravel\Passport\HasApiTokens;
 
 class User extends Authenticatable
 {
@@ -66,6 +68,10 @@ class User extends Authenticatable
     {
         return $this->firstname." ".$this->name;
     }
+
+    protected $casts = [
+        'email_verified_at' => 'datetime',
+    ];
 
 
 }
