@@ -1,15 +1,18 @@
 <template>
     <div class="container">
-        
         <form @submit.prevent="submit" method="POST" enctype="multipart/form-data">
             <!-- the form items -->
-            <global-input type='text' inputName="name" inputId="name" tekstLabel="Naam: " v-model="fields.name" :errors="errors.name" :value='fields.name'></global-input>
-            <!-- Admin multiselect -->            
+             <global-layout>
+                <label>Naam: </label>
+                <input type="text" class="form-control" v-model="fields.name"/>
+                <div class="text-danger" v-if="errors">{{ errors.name }}</div>
+            </global-layout>
+            <!-- Admin multiselect -->    
             <div class="row" v-if="multiGroupUsers">
                 <div class="col-lg-2 col-md-2 col-sm-0"></div>
                 <div class="col-lg-8 col-md-8 col-sm-12">
                     <label>Admin: </label>
-                    <multiselect
+                    <Multiselect
                         v-model="selectedUser"
                         :multiple="false"
                         :options="multiGroupUsers"
@@ -18,17 +21,14 @@
                         placeholder="Select admin"
                         label="full_name"
                         track-by="full_name"
-                    >
-                    </multiselect>
+                    />
                 </div>
                 <div class="col-lg-2 col-md-2 col-sm-0"></div>
             </div>
-            
+
             <br>
-            <global-layout>
-                 <center>
-                    <button class="btn btn-primary">{{buttonText}}</button>
-                </center>
+            <global-layout center="center">
+                <button class="btn btn-primary">{{buttonText}}</button>
                 <br>
             </global-layout>
         </form>
@@ -37,15 +37,10 @@
 
 <script>
     import apiCall from '../../services/ApiCall.js';
-
-    import TextInput from '../../components/ui/form/TextInput.vue';
-    import TextArea from '../../components/ui/form/TextAreaInput.vue';
-    import Multiselect from 'vue-multiselect';
-
+    import Multiselect from '@suadelabs/vue3-multiselect';
+    
     export default {
         components: {
-            TextInput,
-            TextArea,
             Multiselect
         },
 
@@ -53,11 +48,10 @@
             return {
                 'fields' : {
                     name: '',
-                    admin_id: '',
                 },
                 'errors' : {},
-                'selectedUser': [],
-                'multigroupUsers': [],
+                'selectedUser': {},
+                //'multiGroupUsers': [],
                 'formData': new FormData(),
                 'buttonText': "Save group"
             }
@@ -106,9 +100,10 @@
         },
 
         methods: {
+
+
             setFormData(){
                 this.formData.set('name', this.fields.name ?? null);
-                this.formData.append('description', this.fields.description ?? null);
                 this.formData.set('admin_id', this.selectedUser.id ?? null);
             },
 
@@ -130,12 +125,10 @@
 
                 apiCall.postData('group', this.formData)
                 .then(response =>{
-                    this.$bus.$emit('reloadGroupList');
-                    this.$bus.$emit('resetGroupDisplay');
-                    
                     this.resetFormData();
                     var message = "Your group " + response.data.name + " has been created";
-                    this.$bus.$emit('showMessage', message,  'green', '2000' );
+                    this.$store.dispatch('getMessage', {message: message});
+                    this.$store.dispatch('getUserGroups');
                 }).catch(error => {
                     this.errors = error;
                 });
@@ -146,13 +139,13 @@
 
                 apiCall.updateData('group/' + this.group.id, this.formData)
                 .then(response =>{
-                    this.$bus.$emit('reloadGroup');
                     if(response.data.name == undefined){
-                        this.$bus.$emit('showMessage', response.data,  'red', '2000' );
+                        this.$store.dispatch('getMessage', {message: response.data, color: 'red', time: 2000});
                     }else{
                         var message = "Your group " + response.data.name + " has been changed";
-                        this.$bus.$emit('showMessage', message,  'green', '2000' );
+                        this.$store.dispatch('getMessage', {message: message});
                     }
+                    this.$store.dispatch('getSelectedGroup', {id: this.group.id});
                     this.$router.push({name: "groupDetail", params: { id: this.group.id },})
                 }).catch(error => {
                         this.errors = error;
