@@ -9,7 +9,7 @@ use Tests\TestCase;
 use App\Models\User;
 
 use Database\Seeders\GeneratorSeeder;
-use Laravel\Passport\Passport;
+use Laravel\Sanctum\Sanctum;
 
 class GroupTest extends TestCase
 {
@@ -34,7 +34,7 @@ class GroupTest extends TestCase
      *  Get authenticated user
      */
     protected function authenticatedUser($role = "Admin"){
-        $user = Passport::actingAs(
+        $user = Sanctum::actingAs(
             $this->allUsers[0],
             ['create-servers']
         );
